Add Florist link to .lycheeignore
This has been failing in the link checker, but always work by hand. It should be checked from time to time by hand, because it could disappear at any time (professor emeritus pages). In that case, it could be changed by one of the GitHub repositories (AdaCore/florist, Blady-Com/florist).